Improved LPT dummy feature
It is now based on the development carried out in the 86Box emulator when the selected device is 'None':

https://github.com/86Box/86Box/blob/master/src/lpt.c

With this, pressing the '*' key on the numeric keypad without KEYB in memory on the Tandy 1000 no longer causes problems with the PRN device.

Thank you @somhi for alerting me to this bug.