XTIDE ROM update
The "Full Operation Mode" option is now available in the ".\SW\ROMs\BASICA folder", in the default ".\SW\ROMs" folder the previous version is kept, after it was detected that XTIDE's "Full Operation Mode" option generates problems with several games in Tandy mode. Thanks akeley for report it.

Therefore, it is now up to the user to decide which ROM to use, depending on whether he/she is interested in BASICA or not.