Release 20230312
* CGA CRTC updated to UM6845R
* Hercules CRTC updated to UM6845R

Several glitches of AREA5150 have been solved. The only glitch is the final part of the demo, the water effect.

There are also some minor issues with the edges that have yet to be resolved, which are more evident in this demo, but overall it looks very good.

Thank you @gyurco for the idea and the tips for implementation.